Modifying tests for dynamodb streams
